Match results front end components: standing and report-score and backend changes to store match results.  Created notification to inform players when a match result is reported.

+++ demo/ladderbackend/src/main/java/com/example/demo/dao/MatchDao.java
@@ -5,8 +5,8 @@ import com.example.demo.model.Match;
 
 public interface MatchDao {
 
-    List<Match> getScheduledMatches(Integer ladderId);
-    Integer addMatch(Match match);
+    List<Match> getScheduledMatches(Long ladderId);
+    Long addMatch(Match match);
     void updateMatch(Match match);  
 
 }
